+++ SmartHome/SmartHome/BusinessEntities/BusinessHomeDevice/BusinessLight.cs
@@ -10,25 +10,25 @@ namespace SmartHome.BusinessEntities.BusinessHomeDevice
         [OperationAttribute]
         public static OperationMessage On(this Light light)
         {
-            return OperationMessage.LogicWrite(light.Id, LogicWriteValues.Set, 0);
+            return OperationMessage.LogicWrite((ushort)light.Id, LogicWriteValues.Set, 0);
         }
 
         [OperationAttribute]
         public static OperationMessage Off(this Light light)
         {
-            return OperationMessage.LogicWrite(light.Id, LogicWriteValues.Clear, 0);
+            return OperationMessage.LogicWrite((ushort)light.Id, LogicWriteValues.Clear, 0);
         }
 
         [OperationAttribute]
         public static OperationMessage OnTime(this Light light, byte seconds)
         {
-            return OperationMessage.LogicWrite(light.Id, LogicWriteValues.Set, seconds);
+            return OperationMessage.LogicWrite((ushort)light.Id, LogicWriteValues.Set, seconds);
         }
 
         [OperationAttribute]
         public static OperationMessage Switch(this Light light)
         {
-            return OperationMessage.LogicSwitch(light.Id, 0);
+            return OperationMessage.LogicSwitch((ushort)light.Id, 0);
         }
     }
 }
